backend/cow_identify/requirements.txt
flask
flask-cors
ultralytics
opencv-python
numpy
pillow
